sunxi: Fix PRCM PLL VDO voltage setting
The PRCM register description in various manuals gives two tables to
select the PLL voltage. Since we leave the table select bit (bit 15) as 0,
we are using the table with the higher voltages, so we need to use that
other macro and a different voltage.
This actually doesn't change the register, as the encoding is the same,
but the code reads correctly now and matches the manual, which says that:
"... PLLVDD LDO should be set to 1.37v".

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
Signed-off-by: Samuel Holland <samuel@sholland.org>